requirements.txt
Flask==2.0.1
boto3==1.18.0
python-dotenv==0.19.0
Werkzeug==2.0.3
